added css for resume review and moved functions to utils folder
