+++ packages/frontend/app/app.css
@@ -31,3 +31,58 @@ body {
     color-scheme: dark;
   }
 }
+
+/* CUSTOM EDITOR CLASSES */
+
+/* editor-styles.css */
+.editor-paragraph {
+  margin: 0 0 8px 0;
+  position: relative;
+}
+
+/* Resume-specific styles */
+.resume-name {
+  font-size: 22px;
+  font-weight: bold;
+  text-align: center;
+  margin: 0 0 4px 0;
+}
+
+.resume-contact-info {
+  font-size: 12px;
+  text-align: center;
+  margin: 0 0 12px 0;
+}
+
+.resume-section-header {
+  font-size: 14px;
+  font-weight: bold;
+  margin: 10px 0 5px 0;
+  border-bottom: 1px solid #ddd;
+  padding-bottom: 2px;
+}
+
+.resume-job-title {
+  font-size: 12px;
+  font-weight: bold;
+  margin: 8px 0 2px 0;
+}
+
+.resume-job-date {
+  font-size: 12px;
+  font-style: italic;
+  margin: 0 0 4px 0;
+}
+
+.resume-bullet-point {
+  margin: 2px 0;
+  padding-left: 20px;
+  position: relative;
+}
+
+.resume-bullet-point::before {
+  content: "•";
+  position: absolute;
+  left: 6px;
+}
+
