+++ packages/frontend/package.json
@@ -11,11 +11,18 @@
   "dependencies": {
     "@ai-sdk/google": "^1.1.14",
     "@ai-sdk/openai": "^1.1.12",
+    "@lexical/html": "^0.25.0",
+    "@lexical/react": "^0.25.0",
+    "@lexical/rich-text": "^0.25.0",
     "@react-router/node": "^7.2.0",
     "@react-router/serve": "^7.2.0",
     "ai": "^4.1.41",
     "bad-words": "^4.0.0",
+    "docx": "^9.2.0",
+    "file-saver": "^2.0.5",
     "isbot": "^5.1.22",
+    "jspdf": "^3.0.0",
+    "lexical": "^0.25.0",
     "openai": "^4.85.2",
     "react": "19.0.0",
     "react-dom": "19.0.0",
